ml-solutions/requirements.txt
black
pylint
joblib
scikit-learn
google-cloud-vision
google-cloud-translate
google-cloud-aiplatform
